requirements.txt
python-dotenv>=1.0.0
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
mysql-connector-python
schedule==1.2.0 